Add files via upload
Number Guessing Game
This guess the number game is a short TypeScript/Node.js project that allows the user to guess the number generated by the computer. There are also several ways to alter the game, like adding more rounds or displaying the score. It’s quite simple and uses the random function to generate a number.

Create a GitHub repository for the project and submit its URL in the project submission form